tmp fix for failing tests

+++ pinot-core/src/test/java/org/apache/pinot/core/operator/transform/function/TupleSelectionTransformFunctionsTest.java
@@ -66,7 +66,7 @@ public class TupleSelectionTransformFunctionsTest extends BaseTransformFunctionT
   public void testLeastTransformFunctionInt() {
     TransformFunction transformFunction = testLeastPreconditions(
         String.format("least(%s, %d, cast(%s as INT))", INT_SV_COLUMN, -1, FLOAT_SV_COLUMN));
-    assertEquals(transformFunction.getResultMetadata().getDataType(), FieldSpec.DataType.INT);
+    assertEquals(transformFunction.getResultMetadata().getDataType(), FieldSpec.DataType.LONG);
     int[] intValues = transformFunction.transformToIntValuesSV(_projectionBlock);
     for (int i = 0; i < NUM_ROWS; i++) {
       assertEquals(intValues[i], Math.min(Math.min(_intSVValues[i], -1), (int) _floatSVValues[i]));
